pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = "Test task for career center"
authors = ["ageht <Vladimir.E.Pavlov@yandex.ru>"]
readme = "README.md"
packages = [{include = "test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = true
skip-magic-trailing-coma = true
exclude = '''
(
  /(
      \.git
    | \.github
    | \.venv
  )/
)
'''
